hinagata更新 mint作成

+++ hinagata.cpp
@@ -1,8 +1,8 @@
 #include <bits/stdc++.h>
-#define rep(i,n)    for(int i = 0; i < (n) ; ++i)
-#define orep(i,n)   for(int i = 1; i <= (n); ++i)   // one rep
-#define drep(i,n)   for(int i = (n)-1; i >= 0; --i) // down rep
-#define srep(i,s,t) for(int i = (s); i < (t); ++i)  // set rep
+#define rep(i,n)    for(long long i = 0; i < (n) ; ++i)
+#define orep(i,n)   for(long long i = 1; i <= (n); ++i)   // one rep
+#define drep(i,n)   for(long long i = (n)-1; i >= 0; --i) // down rep
+#define srep(i,s,t) for(long long i = (s); i < (t); ++i)  // set rep
 #define rng(x)  (x).begin(),(x).end()   // range
 #define rrng(a) (a).rbegin(),(a).rend() // reverse range
 #define isin(x,l,r) ((l) <= (x) && (x) < (r))
